+++ services/core/shared/core/compat/sdk_v4/dynamicFees.js
@@ -13,6 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
+const BluebirdPromise = require('bluebird');
 const { CacheRedis } = require('lisk-service-framework');
 
 const {
@@ -24,8 +25,15 @@ const {
 	ProofOfMisbehaviorTransaction,
 } = require('@liskhq/lisk-transactions-v4');
 
-const config = require('../../../../config');
+const { calcAvgFeeByteModes, EMAcalc } = require('../common/dynamicFees');
+const { mapToOriginal } = require('./reverseMappings');
+const { getBlocks } = require('./blocks');
+const { getTransactions } = require('./transactions');
 
+const config = require('../../../../config');
+const requestAll = require('../../../requestAll');
+
+const cacheRedisFees = CacheRedis('fees', config.endpoints.redis);
 const cacheRedisBlockSizes = CacheRedis('blockSizes', config.endpoints.redis);
 
 const getTransactionInstanceByType = transaction => {
@@ -62,7 +70,168 @@ const calculateBlockSize = async block => {
 	return blockSize;
 };
 
+const calculateWeightedAvg = async blocks => {
+	const blockSizes = await Promise.all(blocks.map(block => calculateBlockSize(block)));
+	const decayFactor = config.feeEstimates.wavgDecayPercentage / 100;
+	let weight = 1;
+	const wavgLastBlocks = blockSizes.reduce((a = 0, b = 0) => {
+		weight *= 1 - decayFactor;
+		return a + (b * weight);
+	});
+
+	return wavgLastBlocks;
+};
+
+const calculateAvgFeePerByte = (mode, transactionDetails) => {
+	const maxBlockSize = 15 * 2 ** 10;
+	const allowedModes = Object.values(calcAvgFeeByteModes);
+
+	const lowerPercentile = allowedModes.includes(mode) && mode === calcAvgFeeByteModes.MEDIUM
+		? config.feeEstimates.medEstLowerPercentile : config.feeEstimates.highEstLowerPercentile;
+	const upperPercentile = allowedModes.includes(mode) && mode === calcAvgFeeByteModes.MEDIUM
+		? config.feeEstimates.medEstUpperPercentile : config.feeEstimates.highEstUpperPercentile;
+	const lowerBytePos = Math.ceil((lowerPercentile / 100) * maxBlockSize);
+	const upperBytePos = Math.floor((upperPercentile / 100) * maxBlockSize);
+
+	let currentBytePos = 0;
+	let totalFeePriority = 0;
+	transactionDetails.forEach(transaction => {
+		if (currentBytePos <= lowerBytePos && lowerBytePos < currentBytePos + transaction.size
+			&& currentBytePos + transaction.size <= upperBytePos) {
+			totalFeePriority += transaction.feePriority
+				* (currentBytePos + transaction.size - lowerBytePos + 1);
+		}
+
+		if (lowerBytePos <= currentBytePos && currentBytePos + transaction.size <= upperBytePos) {
+			totalFeePriority += transaction.feePriority * transaction.size;
+		}
+
+		if (lowerBytePos <= currentBytePos && upperBytePos >= currentBytePos
+			&& upperBytePos <= currentBytePos + transaction.size) {
+			totalFeePriority += transaction.feePriority * (upperBytePos - currentBytePos + 1);
+		}
+
+		currentBytePos += transaction.size;
+	});
+
+	const avgFeePriority = totalFeePriority / (upperBytePos - lowerBytePos + 1);
+	return avgFeePriority;
+};
+
+const calculateFeePerByte = async block => {
+	const feePerByte = {};
+	const transactionDetails = block.payload.map(transaction => {
+		const tx = getTransactionInstanceByType(transaction);
+		const transactionSize = tx.getBytes().length;
+		const { minFee } = tx;
+		const feePriority = (Number(transaction.fee) - Number(minFee)) / transactionSize;
+		return {
+			id: transaction.id,
+			size: transactionSize,
+			feePriority,
+		};
+	});
+	transactionDetails.sort((t1, t2) => t1.feePriority - t2.feePriority);
+
+	const blockSize = await calculateBlockSize(block);
+
+	feePerByte.low = (blockSize < 12.5 * 2 ** 10) ? 0 : transactionDetails[0].feePriority;
+	feePerByte.med = calculateAvgFeePerByte(calcAvgFeeByteModes.MEDIUM, transactionDetails);
+	feePerByte.high = Math.max(calculateAvgFeePerByte(calcAvgFeeByteModes.HIGH, transactionDetails),
+		(1.3 * feePerByte.med + 1));
+
+	return feePerByte;
+};
+
+const getEstimateFeeByteForBlock = async (blockBatch, innerPrevFeeEstPerByte) => {
+	const wavgBlockBatch = await calculateWeightedAvg(blockBatch.data);
+	const sizeLastBlock = await calculateBlockSize(blockBatch.data[0]);
+	const feePerByte = await calculateFeePerByte(blockBatch.data[0]);
+	const feeEstPerByte = {};
+
+	if (wavgBlockBatch > (12.5 * 2 ** 10) || sizeLastBlock > (14.8 * 2 ** 10)) {
+		const EMAoutput = EMAcalc(feePerByte, innerPrevFeeEstPerByte);
+
+		feeEstPerByte.low = EMAoutput.feeEstLow;
+		feeEstPerByte.med = EMAoutput.feeEstMed;
+		feeEstPerByte.high = EMAoutput.feeEstHigh;
+	} else {
+		feeEstPerByte.low = 0;
+		feeEstPerByte.med = 0;
+		feeEstPerByte.high = 0;
+	}
+
+	feeEstPerByte.updated = Math.floor(Date.now() / 1000);
+	feeEstPerByte.blockHeight = blockBatch.data[0].height;
+	feeEstPerByte.blockId = blockBatch.data[0].id;
+
+	return feeEstPerByte;
+};
+
+const getEstimateFeeByteForBatch = async (fromHeight, toHeight, cacheKey) => {
+	// Check if the starting height is permitted by config or adjust acc.
+	fromHeight = config.feeEstimates.defaultStartBlockHeight > fromHeight
+		? config.feeEstimates.defaultStartBlockHeight : fromHeight;
+
+	const cachedFeeEstimate = await cacheRedisFees.get(cacheKey);
+
+	const cachedFeeEstimateHeight = !cacheKey.includes('Quick') && cachedFeeEstimate
+		? cachedFeeEstimate.blockHeight : 0; // 0 implies does not exist
+
+	const prevFeeEstPerByte = fromHeight > cachedFeeEstimateHeight
+		? { blockHeight: fromHeight - 1 } : cachedFeeEstimate;
+
+	const range = size => Array(size).fill().map((_, index) => index);
+	const feeEstPerByte = {};
+	const blockBatch = {};
+	do {
+		/* eslint-disable no-await-in-loop */
+		const idealEMABatchSize = config.feeEstimates.emaBatchSize;
+		const finalEMABatchSize = idealEMABatchSize > prevFeeEstPerByte.blockHeight
+			? (prevFeeEstPerByte.blockHeight + 1) : idealEMABatchSize;
+
+		blockBatch.data = await BluebirdPromise.map(
+			range(finalEMABatchSize),
+			async i => (await getBlocks({ height: prevFeeEstPerByte.blockHeight + 1 - i })).data[0],
+			{ concurrency: finalEMABatchSize },
+		);
+
+		blockBatch.data = await BluebirdPromise.map(
+			blockBatch.data,
+			async block => {
+				const payload = mapToOriginal(await requestAll(getTransactions, { blockId: block.id }), 'transactions');
+				Object.assign(block, { payload });
+				return block;
+			},
+			{ concurrency: blockBatch.data.length },
+		);
+
+		Object.assign(prevFeeEstPerByte,
+			await getEstimateFeeByteForBlock(blockBatch, prevFeeEstPerByte));
+
+		// Store intermediate values, in case of a long running loop
+		if (prevFeeEstPerByte.blockHeight < toHeight) {
+			await cacheRedisFees.set(cacheKey, prevFeeEstPerByte);
+		}
+
+		/* eslint-enable no-await-in-loop */
+	} while (toHeight > prevFeeEstPerByte.blockHeight);
+
+	Object.assign(feeEstPerByte, prevFeeEstPerByte);
+	await cacheRedisFees.set(cacheKey, feeEstPerByte);
+
+	return feeEstPerByte;
+};
+
 module.exports = {
+	getEstimateFeeByteForBatch,
+
+	// For unit tests
 	getTransactionInstanceByType,
 	calculateBlockSize,
+	calculateWeightedAvg,
+	calculateAvgFeePerByte,
+	calculateFeePerByte,
+	getEstimateFeeByteForBlock,
+	EMAcalc,
 };
